Rename Provider component
